+++ src/components/editor/header/main-toolbar.tsx
@@ -12,8 +12,8 @@ import { TextAlignButton } from "@/components/tiptap/tiptap-ui/text-align-button
 import { UndoRedoButton } from "@/components/tiptap/tiptap-ui/undo-redo-button";
 import { EditorContext } from "@/contexts/editor-provider";
 import { useContext } from "react";
-import FontSizeInput from "../tiptap/tiptap-ui/font-size-input/font-size-input";
-import ShareButton from "./header/share-btn";
+import FontSizeInput from "../../tiptap/tiptap-ui/font-size-input/font-size-input";
+import ShareButton from "./share-btn";
 
 export default function MainToolbarContent() {
   const { editor } = useContext(EditorContext);
